# ollantaweb/docs/swagger.yaml
basePath: /
definitions:
...
  api.loginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/api.userView'
    type: object
...
  api.refreshResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
    type: object
...
  api.userView:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      login:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
...
info:
  contact: {}
  description: Ollanta REST API documentation.
  title: Ollanta API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with username and password
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          properties:
            login:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.loginResponse'
      summary: Login
      tags:
      - auth
...
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a refresh token for a new access token
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.refreshResponse'
      summary: Refresh access token
      tags:
      - auth
...
  /api/v1/projects/{key}/badge:
    get:
      description: Returns an SVG badge for project quality metrics
      parameters:
      - description: Project key
        in: path
        name: key
        required: true
        type: string
      - description: Metric (gate_status, bugs, vulnerabilities, code_smells, ncloc,
          coverage)
        in: query
        name: metric
        type: string
      produces:
      - image/svg+xml
      responses:
        "200":
          description: SVG badge
          schema:
            type: string
      summary: Quality badge
      tags:
      - badges
...
swagger: "2.0"